Port RTL/MJIT to Windows MINGW.
    * Makefile.in: Use relative path to locate the mjit header file and
    add options to link to libruby at runtime.

    * common.mk: include pragma statements into the mjit header file -
    they are required on MINGW. This is not a complete solution, since
    it leaves unsimmetric push/pop statements.

    * compile.c: Enforce 64-bit type on windows for lindex_t

    * mjit.c: Add Windows specific header files.
    Add simple replacements for dl* and waitpid functions.
    Prefer the diectory set per TMP environment variable to "/tmp" -
    this is usually set on Windows but "/tmp" isn't available.
    Properly escape the include path.
    Add functions required for a Windows DLL when built per -nostdlib.
    Use spawn() instead of fork() on Windows.
    Link to ruby.dll.a, because Windows DLLs don't allow unresolved symbols.
    Check addr for NULL because function symbols will never be NULL
    and to make it Windows compatible.
    Use only builtin relative paths and use Ruby's base directory instead.

    * rtl_exec.c: fix method signatures to match insns.def

    * ruby.c: Move mjit_init() after ruby base directory detection.

    * tool/minimize_mjit_header.rb: Use Ruby's cross platform null device.

    * export rb_vm_ep_in_heap_p() for use in VM_ASSERT.

    * vm_core.h: Enforce 64 bit unsigned integer for lindex_t.

    * Export rb_threadptr_* functions for use in inline functions.

+++ compile.c
@@ -198,7 +198,7 @@ r_value(VALUE value)
 #define USE_SELF_LD 0
 
 /* Different macros to convert to/from lindex_t.  */
-#define INT2LINT(x) ((long) (x))
+#define INT2LINT(x) ((ssize_t) (x))
 #define LINT2INT(x) ((int) (x))
 #define FIXNUM2LINT(x) ((int) (x))
